Cleaned up compiler warning messages about doubles being converted to float may cause data loss.

+++ raytrace/Sphere.cpp
@@ -60,8 +60,8 @@ void Sphere::GetSphereUV(const Vec3& p, float& u, float &v)
 {
 	float phi = atan2(p.z(), p.x());
 	float theta = asin(p.y());
-	u = 1 - (phi + M_PI) / (2 * M_PI);
-	v = (theta + M_PI / 2) / M_PI;
+	u = float(1.0f - (phi + M_PI) / (2.0f * M_PI));
+	v = float((theta + M_PI / 2.0f) / M_PI);
 }
 
 AABB Sphere::GetAABB() const
